Rename build to build_docs.
